Sirius.appimg/app/requirements.txt
astroid==2.0.4
autopep8==1.4
beautifulsoup4==4.6.3
boto==2.49.0
boto3==1.9.10
botocore==1.12.10
bs4==0.0.1
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-common-helpers==0.9.2
django-cors-headers==2.3.0
django-cron==0.5.1
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
mutagen==1.40.0
oauth2client==4.1.3
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==4.0
s3transfer==0.1.13
selenium==3.4.3
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
